% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actf.inv.r
\name{actf.inv}
\alias{actf.inv}
\title{Inverse of ACTF auto-correlation transformation function}
\usage{
actf.inv(rhoz, b, c)
}
\arguments{
\item{rhoz}{marginal correlation value of the parent Gaussian process}

\item{b}{1st line parameter}

\item{c}{2nd line parameter}
}
\description{
Provides inverse transformation for continuous distributions, based on two parameters
}
\examples{

actf.inv(.4, 1, 0)

}
\keyword{internal}
