% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actf.inv.r
\name{actfdiscrete.inv}
\alias{actfdiscrete.inv}
\title{Inverse of ACTF auto-correlation transformation function}
\usage{
actfdiscrete.inv(rhoz, b, c)
}
\arguments{
\item{rhoz}{marginal correlation value of the parent Gaussian process}

\item{b}{1st line parameter}

\item{c}{2nd line parameter}
}
\description{
Provides inverse transformation for continuous distributions, based on two parameters
}
\examples{

actfdiscrete.inv(.4, .2, 1)

}
\keyword{internal}
