% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actf.R
\name{actfdiscrete}
\alias{actfdiscrete}
\title{ACTF auto-correlation transformation function for discrete distributions}
\usage{
actfdiscrete(rhox, b, c)
}
\arguments{
\item{rhox}{marginal correlation value}

\item{b}{1st parameter}

\item{c}{2nd parameter}
}
\value{
numeric vector of transformed Gaussian correlations
}
\description{
Provides transformation for discrete distributions, based on two parameters.
}
\examples{
actfdiscrete(.4, .2, 1)

}
\seealso{
\code{\link{actpnts}}, \code{\link{fitactf}}
}
\keyword{internal}
