% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actf.R
\name{actfInv}
\alias{actfInv}
\title{Inverse ACTF}
\usage{
actfInv(rhoz, b, c)
}
\arguments{
\item{rhoz}{Gaussian correlation value}

\item{b}{1st parameter}

\item{c}{2nd parameter}
}
\value{
numeric vector of back-transformed marginal correlations
}
\description{
Inverse of the autocorrelation transformation function.
}
\examples{
actfInv(.4, 1, 0)

}
\seealso{
\code{\link{actf}}
}
\keyword{internal}
