% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actf.R
\name{actpntsB6}
\alias{actpntsB6}
\title{AutoCorrelation Transformed Points for Bardossy dependence structure}
\usage{
actpntsB6(margdist, margarg, m, p0 = 0)
}
\arguments{
\item{margdist}{target marginal distribution}

\item{margarg}{list of marginal distribution arguments}

\item{m}{mean of the parent Gaussian processes controlling asymmetry}

\item{p0}{probability of zero values}
}
\value{
A data frame with columns \code{rhoz} and \code{rhox}.
}
\description{
Evaluates the (rho_x, rho_z) mapping using Monte Carlo integration for the
Bardossy copula dependence structure.
}
\examples{

library(CoSMoS)

x <- actpntsB6(margdist = "paretoII",
               margarg  = list(scale = 1, shape = .3),
               m = 1,
               p0 = 0)
x

}
\references{
Bardossy, A. (2006). Copula-based geostatistical models for
groundwater quality parameters. Water Resources Research, 42(11),
\doi{10.1029/2005WR004754}
}
\seealso{
\code{\link{actpnts}}, \code{\link{generateMTSFast}}
}
