% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advectionF.R
\name{advectionF2}
\alias{advectionF2}
\title{Advection fields}
\usage{
advectionF2(id, arglist)
}
\arguments{
\item{id}{advection type id}

\item{arglist}{list of additional arguments (vector of coordinates and parameters of advection field functions)}
}
\description{
Provides parametric functions that describe different types of advection fields.
}
\references{
Papalexiou, S.M., Serinaldi, F., Porcu, E. (2021). Advancing
Space-Time Simulation of Random Fields: From Storms to Cyclones and Beyond.
Water Resources Research, 57, e2020WR029466, \doi{10.1029/2020WR029466}
}
\keyword{internal}
