% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anisotropyT.R
\name{anisotropyT2}
\alias{anisotropyT2}
\title{Anisotropy transformation}
\usage{
anisotropyT2(id, arglist)
}
\arguments{
\item{id}{anisotropy type id}

\item{arglist}{list of additional arguments (vector of coordinates and parameters of the anisotropy transformations)}
}
\description{
Provides parametric functions that describe different types of planar deformation fields, including affine (rotation and stretching), and swirl-like deformation. For more details see Papalexiou et al.(2021) and references therein.
}
\references{
Papalexiou, S.M., Serinaldi, F., Porcu, E. (2021). Advancing
Space-Time Simulation of Random Fields: From Storms to Cyclones and Beyond.
Water Resources Research, 57, e2020WR029466, \doi{10.1029/2020WR029466}
}
\keyword{internal}
