% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{BurrXII}
\alias{BurrXII}
\alias{dburrXII}
\alias{pburrXII}
\alias{qburrXII}
\alias{rburrXII}
\alias{mburrXII}
\title{Burr Type XII distribution}
\usage{
dburrXII(x, scale, shape1, shape2, log = FALSE)

pburrXII(q, scale, shape1, shape2, lower.tail = TRUE, log.p = FALSE)

qburrXII(p, scale, shape1, shape2, lower.tail = TRUE, log.p = FALSE)

rburrXII(n, scale, shape1, shape2)

mburrXII(r, scale, shape1, shape2)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{scale, shape1, shape2}{scale and shape parameters; the shape
arguments cannot be vectors (must have length one).}

\item{log, log.p}{logical; if \code{TRUE}, probabilities \code{p} are
given as \code{log(p)}.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are
\eqn{P[X \le x]}, otherwise \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length
is taken to be the number required.}

\item{r}{raw moment order.}
}
\value{
\code{dburrXII} returns a numeric vector of density values.
\code{pburrXII} returns a numeric vector of cumulative probabilities.
\code{qburrXII} returns a numeric vector of quantiles.
\code{rburrXII} returns a numeric vector of random deviates.
\code{mburrXII} returns the raw moment of order \code{r}.
}
\description{
Provides density, distribution function, quantile function, random value
generation, and raw moments of order \emph{r} for the Burr Type XII
distribution.
}
\examples{

## plot the density

ggplot(data.frame(x = c(0, 10)),
       aes(x)) +
  stat_function(fun = dburrXII,
                args = list(scale = 5,
                            shape1 = .25,
                            shape2 = .75),
                colour = "royalblue4") +
  labs(x = "",
       y = "Density") +
  theme_classic()
}
\references{
Papalexiou, S.M. (2018). Unified theory for stochastic
modelling of hydroclimatic processes: Preserving marginal distributions,
correlation structures, and intermittency. Advances in Water Resources,
115, 234-252, \doi{10.1016/j.advwatres.2018.02.013}
}
\seealso{
\code{\link{fitDist}}, \code{\link{moments}}
}
\concept{Continuous}
\concept{Univariate}
\keyword{distribution}
