% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkplot}
\alias{checkplot}
\title{Plot function for check results}
\usage{
checkplot(x, margdist, margarg, p0)
}
\arguments{
\item{x}{check result}
}
\description{
Plot function for check results
}
\keyword{internal}
