% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSTRF.r
\name{checkSTRF}
\alias{checkSTRF}
\title{Numerical and visual check of generated random fields}
\usage{
checkSTRF(STRF, lags = 30)
}
\arguments{
\item{STRF}{output of 'generateSTRF' or 'generateSTRFsepfast'}

\item{lags}{number of lags of empirical STCF to be considered in the graphical output (default set to 30)}
}
\description{
Compares generated random fields sample statistics with the theoretically expected values (similar to checkTS). It also returns graphical output for visual check
}
\examples{

fit <- fitVARSTRF(
  m = 10,
  p = 1,
  margdist ='burrXII',
  margarg = list(scale = 3, shape1 = .9, shape2 = .2),
  p0 = 0.8,
  stcsid = "clayton",
  stcsarg = list(scfid = "weibull", tcfid = "weibull",
                 copulaarg = 2,
                 scfarg = list(scale = 20, shape = 0.7),
                tcfarg = list(scale = 1.1, shape = 0.8))
)

sim <- generateSTRF(n = 10,
                    STmodel = fit)
checkSTRF(STRF = sim,
          lags = 10)


}
