% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disch.R
\docType{data}
\name{disch}
\alias{disch}
\title{Daily streamflow data data}
\format{
A data.table with 23315 rows and 2 variables:
\describe{
\item{date}{POSIXct format date/time}
\item{value}{daily avarage values}
}
}
\source{
The United States Geological Survey (USGS) National Water Information System (NWIS)
}
\usage{
disch
}
\description{
Station details
\itemize{
\item Name: Nassawango Creek near Snow Hill, Worcester County, Maryland, Hydrologic Unit 02080111
\item Network Id: , USGS 01485500
\item Latitude/Longitude: 38°13'44.1", 75°28'17.2"
\item Elevation: 11.49 ft above North American Vertical Datum of 1988.
\item Measurement unit: cubic feet per second
}
}
\details{
more details can be found \href{https://waterdata.usgs.gov/nwis/dv?referred_module=sw&site_no=01485500}{here}.
}
\keyword{datasets}
