% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{erfc}
\alias{erfc}
\alias{inv.erfc}
\title{Complementary error function}
\usage{
erfc(x)

inv.erfc(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector
}
\description{
Complementary error function

Inverse complementary error function
}
\keyword{internal}
