% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{rMSE}
\alias{rMSE}
\alias{MSE}
\title{Ratio mean square error}
\usage{
rMSE(x, y)

MSE(x, y)
}
\arguments{
\item{x}{vector of observed values}

\item{y}{vector of simulated values}
}
\description{
Ratio mean square error
}
\examples{

rMSE(rnorm(10), rnorm(10))

MSE(rnorm(10), rnorm(10))

}
\keyword{internal}
