% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitVARSTRF.r
\name{fitVARSTRF}
\alias{fitVARSTRF}
\title{VAR model parameters to simulate parent Gaussian random fields}
\usage{
fitVARSTRF(
  m,
  p,
  margdist,
  margarg,
  p0,
  distbounds = c(-Inf, Inf),
  stcsid,
  stcsarg,
  scalefactor = 1,
  anisotropyarg = list(phi1 = 1, phi2 = 1, theta = 0)
)
}
\arguments{
\item{m}{side length of the square field (m x m)}

\item{p}{order of VAR(p) model}

\item{margdist}{target marginal distribution of the field}

\item{margarg}{list of marginal distribution arguments}

\item{p0}{probability zero}

\item{distbounds}{distribution bounds (default set to c(-Inf, Inf))}

\item{stcsid}{spatiotemporal correlation structure ID}

\item{stcsarg}{list of spatiotemporal correlation structure arguments}

\item{scalefactor}{factor specifying the distance between the centers of two pixels (default set to 1)}

\item{anisotropyarg}{list of arguments establishing spatial anisotropy. phi1 and phi2 control the stretch in two orthogonal directions (e.g., longitude and latitude) while the angle theta controls a counterclockwise rotation (default set to list(phi1 = 1, phi2 = 1 , theta = 0) for isotropic fields)}
}
\description{
Compute VAR model parameters to simulate parent Gaussian random fields with specified spatiotemporal correlation structure using the method described by \href{https://doi.org/10.1145/937332.937333}{Biller and Nelson (2003)}
}
\examples{

fit <- fitVARSTRF(
  m = 10,
  p = 1,
  margdist ='burrXII',
  margarg = list(scale = 3, shape1 = .9, shape2 = .2),
  p0 = 0.8,
  stcsid = "clayton",
  stcsarg = list(scfid = "weibull", tcfid = "weibull",
                 copulaarg = 2,
                 scfarg = list(scale = 20, shape = 0.7),
                tcfarg = list(scale = 1.1, shape = 0.8))
)

dim(fit$alpha)
dim(fit$res.cov)

fit$m
fit$margarg
fit$margdist

}
