% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMULTI.r
\name{generateMULTI}
\alias{generateMULTI}
\title{Simulation of spatiotemporal random vectors with given properties}
\usage{
generateMULTI(n, STmodel)
}
\arguments{
\item{n}{number of fields (time steps) to simulate}

\item{STmodel}{list of arguments resulting from 'fitVARMULTI' function}
}
\description{
Generates spatiotemporal random vectors with given properties, just provide (1) the output of 'fitVARMULTI' function, and (2) the number of time steps to simulate
}
\examples{

coord <- cbind(runif(4)*30, runif(4)*30)

fit <- fitVARMULTI(
  coord = coord,
  p = 1,
  margdist ='burrXII',
  margarg = list(scale = 3,
                 shape1 = .9,
                 shape2 = .2),
  p0 = 0.8,
  stcsid = "clayton",
  stcsarg = list(scfid = "weibull",
                 tcfid = "weibull",
                 copulaarg = 2,
                 scfarg = list(scale = 20,
                               shape = 0.7),
                 tcfarg = list(scale = 1.1,
                               shape = 0.8))
)

sim <- generateMULTI(n = 100,
                     STmodel = fit)

}
