% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSTRF.r
\name{generateSTRF}
\alias{generateSTRF}
\title{Simulation of spatiotemporal random fields with given properties}
\usage{
generateSTRF(n, STmodel)
}
\arguments{
\item{n}{number of fields (time steps) to simulate}

\item{STmodel}{list of arguments resulting from 'fitVAR' function}
}
\description{
Generates spatiotemporal random fields with given properties, just provide (1) the output of 'fitVARSTRF' function, and (2) the number of fields (time steps) to simulate
}
\examples{

fit <- fitVARSTRF(
  m = 10,
  p = 1,
  margdist ='burrXII',
  margarg = list(scale = 3, shape1 = .9, shape2 = .2),
  p0 = 0.8,
  stcsid = "clayton",
  stcsarg = list(scfid = "weibull", tcfid = "weibull",
                 copulaarg = 2,
                 scfarg = list(scale = 20, shape = 0.7),
                tcfarg = list(scale = 1.1, shape = 0.8))
)

sim <- generateSTRF(n = 10,
                    STmodel = fit)
checkSTRF(sim)

}
