% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{MSE}
\alias{MSE}
\title{Mean squared error}
\usage{
MSE(x, y)
}
\arguments{
\item{x}{numeric vector of observed values}

\item{y}{numeric vector of reference values}
}
\value{
scalar numeric
}
\description{
Mean squared error
}
\seealso{
\code{\link{rMSE}}
}
\keyword{internal}
