% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{ParetoII}
\alias{ParetoII}
\alias{dparetoII}
\alias{pparetoII}
\alias{qparetoII}
\alias{rparetoII}
\alias{mparetoII}
\title{Pareto Type II distribution}
\usage{
dparetoII(x, scale, shape, log = FALSE)

pparetoII(q, scale, shape, lower.tail = TRUE, log.p = FALSE)

qparetoII(p, scale, shape, lower.tail = TRUE, log.p = FALSE)

rparetoII(n, scale, shape)

mparetoII(r, scale, shape)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{scale, shape}{scale and shape parameters; the shape argument cannot
be a vector (must have length one).}

\item{log, log.p}{logical; if \code{TRUE}, probabilities \code{p} are
given as \code{log(p)}.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are
\eqn{P[X \le x]}, otherwise \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length
is taken to be the number required.}

\item{r}{raw moment order.}
}
\value{
\code{dparetoII} returns a numeric vector of density values.
\code{pparetoII} returns a numeric vector of cumulative probabilities.
\code{qparetoII} returns a numeric vector of quantiles.
\code{rparetoII} returns a numeric vector of random deviates.
\code{mparetoII} returns the raw moment of order \code{r}.
}
\description{
Provides density, distribution function, quantile function, random value
generation, and raw moments of order \emph{r} for the Pareto type II
distribution.
}
\examples{

## plot the density

ggplot(data.frame(x = c(0, 20)),
       aes(x)) +
  stat_function(fun = dparetoII,
                args = list(scale = 1,
                            shape = .3),
                colour = "royalblue4") +
  labs(x = "",
       y = "Density") +
  theme_classic()
}
\seealso{
\code{\link{fitDist}}, \code{\link{moments}}
}
\concept{Continuous}
\concept{Univariate}
\keyword{distribution}
