% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.acti}
\alias{plot.acti}
\title{Plot method for \code{acti} objects}
\usage{
\method{plot}{acti}(x, ...)
}
\arguments{
\item{x}{an \code{acti} object returned by \code{\link{fitactf}}}

\item{...}{optional arguments; \code{main} sets the plot title}
}
\value{
a \code{ggplot} object (invisibly returned; also printed)
}
\description{
Visualises the autocorrelation transformation function (ACTF) fitted by
\code{\link{fitactf}}.
}
\examples{

library(CoSMoS)

p   <- actpnts(margdist = "paretoII",
               margarg  = list(scale = 1, shape = .3),
               p0 = 0)
fit <- fitactf(p)

plot(fit)
plot(fit, main = "Pareto type II\nautocorrelation transformation")

}
\seealso{
\code{\link{fitactf}}, \code{\link{actpnts}}
}
