% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.checkTS}
\alias{plot.checkTS}
\title{Plot method for \code{checkTS} objects}
\usage{
\method{plot}{checkTS}(x, ...)
}
\arguments{
\item{x}{a \code{checkTS} object returned by \code{\link{checkTS}}}

\item{...}{currently unused}
}
\value{
a \code{ggplot} object (invisibly returned; also printed)
}
\description{
Displays boxplots of simulated statistics against theoretical expected values
for each statistic tracked by \code{\link{checkTS}}.
}
\examples{

library(CoSMoS)

x <- generateTS(margdist = "burrXII",
                margarg = list(scale = 1,
                               shape1 = .75,
                               shape2 = .15),
                acsvalue = acs(id = "weibull",
                               t = 0:30,
                               scale = 10,
                               shape = .75),
                n = 1000, p = 30, p0 = .25, TSn = 100)

chck <- checkTS(x)
plot(chck)

}
\seealso{
\code{\link{checkTS}}
}
