% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.cosmosts}
\alias{plot.cosmosts}
\title{Plot method for \code{cosmosts} objects}
\usage{
\method{plot}{cosmosts}(x, ...)
}
\arguments{
\item{x}{a \code{cosmosts} object returned by \code{\link{generateTS}}}

\item{...}{currently unused}
}
\value{
a \code{ggplot} object (invisibly returned; also printed)
}
\description{
Visualises time series generated by \code{\link{generateTS}} as bar charts,
one panel per series.
}
\examples{

library(CoSMoS)

ts <- generateTS(margdist = "ggamma",
                 margarg  = list(scale = 1, shape1 = .8, shape2 = .8),
                 acsvalue = acs(id = "paretoII", t = 0:30,
                                scale = 1, shape = .75),
                 n = 1000, p = 30, TSn = 2)
plot(ts)

}
\seealso{
\code{\link{generateTS}}, \code{\link{regenerateTS}}
}
