% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.fitACS}
\alias{plot.fitACS}
\title{Plot method for \code{fitACS} objects}
\usage{
\method{plot}{fitACS}(x, ...)
}
\arguments{
\item{x}{a \code{fitACS} object returned by \code{\link{fitACS}}}

\item{...}{currently unused}
}
\value{
a \code{ggplot} object (invisibly returned; also printed)
}
\description{
Displays the empirical ACF alongside the fitted theoretical autocorrelation
structure.
}
\examples{

x <- arima.sim(model = list(ar = 0.8), n = 1000)
acsfit <- fitACS(acf(x, plot = FALSE)$acf, "weibull", c(1, 1))
plot(acsfit)

}
\seealso{
\code{\link{fitACS}}
}
