% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.fitDist}
\alias{plot.fitDist}
\title{Plot method for \code{fitDist} objects}
\usage{
\method{plot}{fitDist}(x, ...)
}
\arguments{
\item{x}{a \code{fitDist} object returned by \code{\link{fitDist}}}

\item{...}{currently unused}
}
\value{
a \code{ggplot} object (invisibly returned; also printed)
}
\description{
Displays the empirical CDF against the fitted theoretical CDF on a
log-exceedance-probability scale.
}
\examples{

x <- fitDist(rnorm(1000), "norm", 30, "N1", FALSE)
plot(x)

}
\seealso{
\code{\link{fitDist}}
}
