% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip.R
\docType{data}
\name{precip}
\alias{precip}
\title{Precipitation data}
\format{A data.table with 79633 rows and 2 variables:
\describe{
\item{date}{POSIXct format date/time}
\item{value}{precipitation totals}
...
}}
\source{
some database
}
\usage{
precip
}
\description{
Precipitation data
}
\keyword{datasets}
