% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{quickTSPlot}
\alias{quickTSPlot}
\title{Quick visualisation of basic time series properties}
\usage{
quickTSPlot(TS, ci = 0.95)
}
\arguments{
\item{TS}{numeric vector (or \code{data.frame}/\code{data.table}) of time
series values}

\item{ci}{numeric; confidence level for the zero-autocorrelation band
(default \code{0.95})}
}
\value{
a \code{ggdraw} object (printed as a side effect)
}
\description{
Returns a composite figure showing the time series, empirical density
function, and empirical autocorrelation function.
}
\examples{
ggamma_sim <- rggamma(n = 1000, scale = 1, shape1 = 1, shape2 = .5)
quickTSPlot(ggamma_sim)

}
\seealso{
\code{\link{generateTS}}, \code{\link{plot.cosmosts}}
}
