% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-ts.R
\name{regenerateTS}
\alias{regenerateTS}
\title{Bulk time series generation}
\usage{
regenerateTS(ts, TSn = 1)
}
\arguments{
\item{ts}{a \code{cosmosts} object returned by \code{\link{generateTS}}}

\item{TSn}{number of time series to generate}
}
\value{
An object of class \code{'cosmosts'}: a list of \code{TSn} numeric
vectors of the same length as those in \code{ts}.
}
\description{
Generates additional time series using parameters already fitted by
\code{\link{generateTS}}, avoiding recomputation of the ACTF.
}
\details{
After calling \code{\link{generateTS}}, use \code{regenerateTS} to generate
more time series with the same fitted parameters. This is faster than
re-running \code{\link{generateTS}} because the ACTF fitting step is skipped.
}
\examples{

library(CoSMoS)

## Fit once
x <- generateTS(margdist = "burrXII",
                margarg = list(scale = 1,
                               shape1 = .75,
                               shape2 = .25),
                acsvalue = acs(id = "weibull",
                               t = 0:30,
                               scale = 10,
                               shape = .75),
                n = 1000, p = 30, p0 = .5, TSn = 3)

## Generate more realisations with the same parameters
r <- regenerateTS(x)

plot(r)

}
\seealso{
\code{\link{generateTS}}, \code{\link{ARp}}
}
