% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{rMSE}
\alias{rMSE}
\title{Ratio mean squared error}
\usage{
rMSE(x, y)
}
\arguments{
\item{x}{numeric vector of observed values}

\item{y}{numeric vector of reference values}
}
\value{
scalar numeric
}
\description{
Ratio mean squared error
}
\seealso{
\code{\link{MSE}}
}
\keyword{internal}
