% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar-models.R
\name{YW}
\alias{YW}
\title{Yule-Walker solver}
\usage{
YW(ACS)
}
\arguments{
\item{ACS}{Numeric vector. ACS values starting from lag 0
(\code{ACS[1] = 1}). The AR order is \code{length(ACS) - 1}.}
}
\value{
Numeric matrix of AR coefficients, shape \code{(p, 1)}.
}
\description{
Solves the Yule-Walker equations to compute AR(\eqn{p}) coefficients from
an autocorrelation structure (ACS) vector. Used internally by
\code{\link{ARp}} and \code{\link{seasonalAR}}.
}
\seealso{
\code{\link{ARp}}, \code{\link{seasonalAR}}
}
\keyword{internal}
