% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\name{ACSfunctions}
\alias{ACSfunctions}
\alias{acfburrXII}
\alias{acfparetoII}
\alias{acffgn}
\alias{acfweibull}
\title{Autocorrelation structure functions}
\usage{
acfburrXII(t, scale, shape1, shape2)

acfparetoII(t, scale, shape)

acffgn(t, H)

acfweibull(t, scale, shape)
}
\description{
Autocorrelation structure functions
}
\examples{

t <- 1

H <- .75
scale <- .2
shape <- .3
shape1 <- .5
shape2 <- .2

acfburrXII(t, scale, shape1, shape2)

acfparetoII(t, scale, shape)

acffgn(t, H)

acfweibull(t, scale, shape)

}
\keyword{internal}
