% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actf.R
\name{actf}
\alias{actf}
\title{ACTF auto-correlation transformation function}
\usage{
actf(rhox, b, c)
}
\arguments{
\item{rhox}{marginal correlation value}

\item{b}{1st line parameter}

\item{c}{2nd line parameter}
}
\description{
Provides tranformation for continuous distributions, based on two parameters.
}
\examples{

actf(.4, 1, 0)

}
\keyword{internal}
