% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actf.R
\name{actfdiscrete}
\alias{actfdiscrete}
\title{ACTF auto-correlation transformation function for discrete distributions}
\usage{
actfdiscrete(rhox, b, c)
}
\arguments{
\item{rhox}{marginal correlation value}

\item{b}{1st line parameter}

\item{c}{2nd line parameter}
}
\description{
ACTF auto-correlation transformation function for discrete distributions
}
\keyword{internal}
