% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acti.R
\name{acti}
\alias{acti}
\title{ACTI - autocorrelation transformation integral function}
\usage{
acti(x, y, dist, distarg, rhoz, p0)
}
\arguments{
\item{x}{x-plain value}

\item{y}{y-plain value}

\item{dist}{distribution}

\item{distarg}{a list of distribution arguments}

\item{rhoz}{Gaussian correlation}

\item{p0}{probability od zero values}
}
\description{
Expression supplied to double integral.
}
\examples{

acti(1, -1, 'norm', list(), .3, 0)

}
\keyword{internal}
