% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burrIII.R
\name{BurrIII}
\alias{BurrIII}
\alias{dburrIII}
\alias{pburrIII}
\alias{qburrIII}
\alias{rburrIII}
\alias{mburrIII}
\title{Burr Type III distribution}
\usage{
dburrIII(x, scale, shape1, shape2, log = FALSE)

pburrIII(q, scale, shape1, shape2, lower.tail = TRUE, log.p = FALSE)

qburrIII(p, scale, shape1, shape2, lower.tail = TRUE, log.p = FALSE)

rburrIII(n, scale, shape1, shape2)

mburrIII(r, scale, shape1, shape2)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{scale, shape1, shape2}{scale and shape parameters; the shape arguments cannot be a vectors (must have length one).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]} otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}

\item{r}{raw moment order}
}
\description{
Density, distribution function, quantile function, random value generation,
and raw moments of order \emph{r} for the Burr Type III distribution.
}
\concept{Continuous}
\concept{Univariate}
\keyword{distribution}
