% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf.R
\name{ECDF}
\alias{ECDF}
\title{Empirical cummulative distrubution function}
\usage{
ECDF(x)
}
\arguments{
\item{x}{vector of values}
}
\description{
Calculates ecdf based on weibull plotting position
}
\examples{

ECDF(round(rnorm(100)))

}
\keyword{internal}
