% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erfc.R
\name{erfc}
\alias{erfc}
\alias{inv.erfc}
\title{Complementary (inverse) error function}
\usage{
erfc(x)

inv.erfc(x)
}
\arguments{
\item{x}{number vector}
}
\description{
Complementary (inverse) error function
}
\examples{

erfc(1)
inv.erfc(1)

}
\keyword{internal}
