% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTS.R
\name{generateTS}
\alias{generateTS}
\title{Generate timeseries}
\usage{
generateTS(margdist, margarg, n, p, p0 = 0, TSn = 1,
  distbounds = c(-Inf, Inf), limitorder = TRUE, acsvalue = NULL)
}
\arguments{
\item{margdist}{target marginal distribution}

\item{margarg}{list of marginal distribution arguments}

\item{n}{number of values}

\item{p}{model order}

\item{p0}{probability zero}

\item{TSn}{number of timeseries to be generated}

\item{distbounds}{distribution bounds (default set to c(-Inf, Inf))}

\item{limitorder}{logical (limit maximum model order according to auto-correlation structure values)}

\item{acsvalue}{target auto-correlation structure (from lag 0)}
}
\description{
Generates timeseries with given properties, just provide (1) the target marginal distribution and its parameters, (2) the target autocorrelation structure or individual autocorrelation values up to a desired lag, and (3) the probablility zero if you wish to simulate an intermittent process.
}
\examples{

library(CoSMoS)

## define target and marginal distributions and arguments
x <- generateTS(margdist = 'burrXII',
                margarg = list(scale = 1,
                               shape1 = .75,
                               shape2 = .25),
                acsvalue = acs(id = 'weibull',
                               t = 0:30,
                               scale = 10,
                               shape = .75),
                n = 1000, p = 30, p0 = .5, TSn = 3)

## see the results
dta <- data.frame(n = seq_along(x[[1]]), do.call(cbind, x))
names(dta)[2:(length(x) + 1)] <- paste('timeseries', 1:length(x), sep = '_')

m.dta <- melt(dta, id.vars = 'n')

ggplot(m.dta, aes(x = n, y = value)) +
  geom_line() +
  labs(x = '') +
  facet_wrap(~variable, ncol = 1) +
  theme_classic()

}
