% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArg.R
\name{getACSArg}
\alias{getACSArg}
\title{Get names of autocorrelation structure (ACS) function arguments}
\usage{
getACSArg(id)
}
\arguments{
\item{id}{ACS id}
}
\description{
Get names of autocorrelation structure (ACS) function arguments
}
\examples{

getACSArg('weibull')

}
\keyword{internal}
