% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArg.R
\name{getDistArg}
\alias{getDistArg}
\title{Get names of distribution function arguments}
\usage{
getDistArg(dist)
}
\arguments{
\item{dist}{distribution name}
}
\description{
Get names of distribution function arguments
}
\examples{

getDistArg('norm')

}
\keyword{internal}
