% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{N}
\alias{N}
\title{Norm - function to be minimized during distrubution fit}
\usage{
N(par, val, dist, norm, n.points)
}
\arguments{
\item{par}{parameter value}

\item{val}{empirical value}

\item{dist}{name of the distribution to be fitted}

\item{norm}{norm used for distribution fitting - id ('N1', 'N2', 'N3', 'N4')}

\item{n.points}{number of points to be subsetted from ecdf}
}
\description{
Norm - function to be minimized during distrubution fit
}
\examples{

N(c(0,1), rnorm(1000), 'norm', 'N2', 30)

}
\keyword{internal}
