% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTS.R
\name{plot.checkTS}
\alias{plot.checkTS}
\title{Plot method for check results}
\usage{
\method{plot}{checkTS}(x, ...)
}
\arguments{
\item{x}{check result}

\item{...}{other args}
}
\description{
Plot method for check results
}
\examples{

library(CoSMoS)

## check your generated timeseries
x <- generateTS(margdist = 'burrXII',
                margarg = list(scale = 1,
                               shape1 = .75,
                               shape2 = .15),
                acsvalue = acs(id = 'weibull',
                               t = 0:30,
                               scale = 10,
                               shape = .75),
                n = 1000, p = 30, p0 = .25, TSn = 100)

chck <- checkTS(x)

plot(chck)

}
