% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcosmosts.R
\name{plot.cosmosts}
\alias{plot.cosmosts}
\title{Plot generated Timeseries}
\usage{
\method{plot}{cosmosts}(x, ...)
}
\arguments{
\item{x}{\code{\link{fitactf}} result object}

\item{...}{other arguments}
}
\description{
Visualizes Timeseries generated by the package CoSMoS.
}
\examples{

library(CoSMoS)

## generate TS
ts <- generateTS(margdist = 'ggamma',
                 margarg = list(scale = 1,
                                shape1 = .8,
                                shape2 = .8),
                 acsvalue = acs(id = 'paretoII',
                                t = 0:30,
                               scale = 1,
                                shape = .75),
                 n = 1000,
                 p = 30,
                 TSn = 2)

## plot the TS
plot(ts)
}
