% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitACS.R
\name{plot.fitACS}
\alias{plot.fitACS}
\title{Plot method for fitACS}
\usage{
\method{plot}{fitACS}(x, ...)
}
\arguments{
\item{x}{fitACS obbject}

\item{...}{other args}
}
\description{
Plot method for fitACS
}
\examples{

x <- AR1(1000, .8)

acsfit <- fitACS(acf(x, plot = FALSE)$acf, 'weibull', c(1, 1))

plot(acsfit)

}
\keyword{internal}
