% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDist.R
\name{plot.fitDist}
\alias{plot.fitDist}
\title{Plot method for fitDist}
\usage{
\method{plot}{fitDist}(x, ...)
}
\arguments{
\item{x}{fitDist object}

\item{...}{other args}
}
\description{
Plot method for fitDist
}
\examples{

x <- fitDist(rnorm(1000), 'norm', 30, 'N1', FALSE)

plot(x)

}
\keyword{internal}
