% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickTSPlot.R
\name{quickTSPlot}
\alias{quickTSPlot}
\title{Quick visualization of basic timeseries properties}
\usage{
quickTSPlot(TS, ci = 0.95)
}
\arguments{
\item{TS}{timeseries to plot}

\item{ci}{confidence interval around the zero autocorrelation value (default set to 0.95, i.e. 95\% CI)}
}
\description{
Return timeseries diagram, empirical density function, and empirical autocorrelation function.
}
\examples{
no <- 1000
ggamma_sim <- rggamma(n = no, scale = 1, shape1 = 1, shape2 = .5)
quickTSPlot(ggamma_sim)

}
