% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalACF.R
\name{seasonalACF}
\alias{seasonalACF}
\title{Calculate seasonal ACF}
\usage{
seasonalACF(TS, season, lag.max = 50)
}
\arguments{
\item{TS}{time series}

\item{season}{name of the season}

\item{lag.max}{max lag for acf}
}
\description{
Calculate seasonal ACF
}
\examples{

x <- data.frame(date = seq(Sys.Date(), by = 'day', length.out = 1000),
                value = rnorm(1000))

sacf <- seasonalACF(x, 'month')

}
\keyword{internal}
