% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalAR.R
\name{seasonalAR}
\alias{seasonalAR}
\title{Seasonal AR model}
\usage{
seasonalAR(x, ACS, season = "month")
}
\arguments{
\item{x}{vector of dates for gaussian process generation}

\item{ACS}{list of ACS for each season}

\item{season}{season name}
}
\description{
Seasonal AR model
}
\examples{

x <- data.frame(date = seq(Sys.Date(), by = 'day', length.out = 1000),
                value = rnorm(1000))

sacf <- seasonalACF(x, 'month')

y <- suppressWarnings((seasonalAR(x$date, sacf)))

}
\keyword{internal}
