% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcs.r
\name{stcs2}
\alias{stcs2}
\title{SpatioTemporal Correlation Structure}
\usage{
stcs2(id, arglist)
}
\arguments{
\item{id}{spatiotemporal correlation structure ID}

\item{arglist}{list of additional arguments (t as time lag, s as spatial lag (distance),  and stcs parameters)}
}
\description{
Provides a parametric function that describes the values of the linear spatiotemporal autocorrelation up to desired lags. For more details on the parametric spatiotemporal correlation structures see section 2.3 and 2.4 in \href{https://doi.org/10.1029/2019WR026331}{Papalexiou and Serinaldi (2020)}
}
\keyword{internal}
