% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalAR.R
\name{YW}
\alias{YW}
\title{Yule-Walker solver}
\usage{
YW(ACS)
}
\arguments{
\item{ACS}{vector of ACS values}
}
\description{
Yule-Walker solver
}
\examples{

YW(rev(exp(seq(-1, 0, .1))))

}
\keyword{internal}
